- fixed sky panning in Duke.
This needs to take the composite texture into account because panning in Build is based on tile size, not map dimension.
It was also redone to use floating point to get rid of the horrible precision of the scrolling effect.

+++ source/games/duke/src/global.cpp
@@ -107,8 +107,8 @@ int animategoal[MAXANIMATES];
 int animatevel[MAXANIMATES];
 int numclouds;								// cloudy skies
 int16_t clouds[256];
-int16_t cloudx;
-int16_t cloudy;
+float cloudx;
+float cloudy;
 int cloudclock;
 int numcyclers;								// sector lighting effects
 int16_t cyclers[MAXCYCLERS][6];
